meta: move Trott to TSC regular member
At the current time, I'm not able to give Node.js the time and attention
it needs and deserves from a voting TSC member. I'm proud of the work
and efforts I've made as a TSC voting member, and I want to leave before
I'm less happy with my efforts. Thanks for all the trust and good will
over the years.

PR-URL: https://github.com/nodejs/node/pull/50297
Reviewed-By: Tierney Cyren <hello@bnb.im>
Reviewed-By: Richard Lau <rlau@redhat.com>